+++ domain/src/main/java/com/emotionstorage/domain/model/MyPage.kt
@@ -2,6 +2,6 @@ package com.emotionstorage.domain.model
 
 data class MyPage(
     val nickname: String,
-    val days: Long,
+    val days: Int,
     val keys: Int,
 )
